Add files via upload
This project allows teachers to add quiz with a  unique code. This code needs to be entered by students to take the quiz. The system evaluates and gives a score as soon as the student takes the quiz. Any user can view the leaderboard of any quiz by entering its code.